+++ iftttie/services/file_.py
@@ -4,22 +4,23 @@ from asyncio import sleep
 from asyncio.queues import Queue
 from datetime import timedelta
 from pathlib import Path
-from typing import Any
+from typing import Any, Optional
 
 from aiohttp import ClientSession
 from loguru import logger
 
 from iftttie.dataclasses_ import Update
-from iftttie.enums import ValueKind
+from iftttie.enums import Unit
 from iftttie.services.base import BaseService
 
 
 class File(BaseService):
-    def __init__(self, path: Path, key: str, interval: timedelta, kind: ValueKind):
+    def __init__(self, path: Path, key: str, interval: timedelta, unit: Unit, title: Optional[str] = None):
         self.path = path
         self.key = key
         self.interval = interval.total_seconds()
-        self.kind = kind
+        self.unit = unit
+        self.title = title
 
     async def run(self, client_session: ClientSession, event_queue: Queue[Update], **kwargs: Any):
         while True:
@@ -27,7 +28,8 @@ class File(BaseService):
                 await event_queue.put(Update(
                     key=f'file:{self.key}',
                     value=self.preprocess_value(self.path.read_text()),
-                    kind=self.kind,
+                    unit=self.unit,
+                    title=self.title,
                 ))
             except IOError as e:
                 logger.error('I/O error in {service}:', service=self)
@@ -43,8 +45,8 @@ class File(BaseService):
 
 
 class FloatValueFile(File):
-    def __init__(self, path: Path, key: str, interval: timedelta, kind: ValueKind, scale=1.0):
-        super().__init__(path, key, interval, kind)
+    def __init__(self, path: Path, key: str, interval: timedelta, unit: Unit, scale=1.0, title: Optional[str] = None):
+        super().__init__(path, key, interval, unit, title)
         self.scale = scale
 
     def preprocess_value(self, value: str):