+++ iftttie/services/nest.py
@@ -3,15 +3,15 @@ from __future__ import annotations
 from asyncio import Queue
 from datetime import datetime
 from typing import Any, Iterable, List, Tuple
-from ujson import loads
 
 from aiohttp import ClientSession
 from aiohttp_sse_client.client import EventSource, MessageEvent
 from loguru import logger
 from multidict import MultiDict
+from ujson import loads
 
 from iftttie.dataclasses_ import Update
-from iftttie.enums import ValueKind
+from iftttie.enums import Unit
 from iftttie.services.base import BaseService
 
 url = 'https://developer-api.nest.com'
@@ -44,23 +44,23 @@ def yield_updates(event: MessageEvent) -> Iterable[Update]:
     devices: dict = data['devices']
 
     yield from yield_devices_updates(data['structures'], 'nest:structure', [
-        ('away', ValueKind.ENUM),
-        ('wwn_security_state', ValueKind.ENUM),
+        ('away', Unit.ENUM, 'Away'),
+        ('wwn_security_state', Unit.ENUM, 'Security State'),
     ])
     yield from yield_devices_updates(devices['cameras'], 'nest:camera', [
-        ('is_streaming', ValueKind.BOOLEAN),
-        ('is_online', ValueKind.BOOLEAN),
-        ('snapshot_url', ValueKind.IMAGE_URL),
+        ('is_streaming', Unit.BOOLEAN, 'Streaming'),
+        ('is_online', Unit.BOOLEAN, 'Online'),
+        ('snapshot_url', Unit.IMAGE_URL, 'Snapshot'),
     ])
     yield from yield_devices_updates(devices['thermostats'], 'nest:thermostat', [
-        ('ambient_temperature_c', ValueKind.CELSIUS),
-        ('humidity', ValueKind.HUMIDITY),
-        ('is_online', ValueKind.BOOLEAN),
-        ('hvac_state', ValueKind.ENUM),
-        ('target_temperature_c', ValueKind.CELSIUS),
+        ('ambient_temperature_c', Unit.CELSIUS, 'Ambient Temperature'),
+        ('humidity', Unit.HUMIDITY, 'Humidity'),
+        ('is_online', Unit.BOOLEAN, 'Online'),
+        ('hvac_state', Unit.ENUM, 'HVAC'),
+        ('target_temperature_c', Unit.CELSIUS, 'Target Temperature'),
     ])
     yield from yield_devices_updates(devices['smoke_co_alarms'], 'nest:smoke_co_alarm', [
-        ('is_online', ValueKind.BOOLEAN),
+        ('is_online', Unit.BOOLEAN, 'Online'),
     ])
 
     for camera_id, camera in devices['cameras'].items():
@@ -69,12 +69,18 @@ def yield_updates(event: MessageEvent) -> Iterable[Update]:
             yield Update(
                 key=f'nest:camera:{camera_id}:last_animated_image_url',
                 value=last_event['animated_image_url'],
-                kind=ValueKind.IMAGE_URL,
+                unit=Unit.IMAGE_URL,
                 timestamp=datetime.strptime(last_event['start_time'], timestamp_format),
+                title=f'{camera["name"]} Last Event',
             )
 
 
-def yield_devices_updates(devices: dict, prefix: str, keys: List[Tuple[str, ValueKind]]) -> Iterable[Update]:
+def yield_devices_updates(devices: dict, prefix: str, keys: List[Tuple[str, Unit, str]]) -> Iterable[Update]:
     for id_, device in devices.items():
-        for key, kind in keys:
-            yield Update(key=f'{prefix}:{id_}:{key}', value=device[key], kind=kind)
+        for key, unit, title in keys:
+            yield Update(
+                key=f'{prefix}:{id_}:{key}',
+                value=device[key],
+                unit=unit,
+                title=f'{device["name"]} {title}',
+            )