+++ iftttie/web.py
@@ -1,17 +1,11 @@
 from __future__ import annotations
 
-from datetime import datetime, timedelta
-from pickle import loads
-from typing import Dict
-
 import aiosqlite
 import pkg_resources
-import pygal
 from aiohttp import web
 from aiohttp_jinja2 import template
 
-from iftttie import constants
-from iftttie.dataclasses_ import Update
+from iftttie.database import select_latest
 
 routes = web.RouteTableDef()
 favicon_body = pkg_resources.resource_string('iftttie', 'static/favicon.png')
@@ -21,13 +15,10 @@ favicon_body = pkg_resources.resource_string('iftttie', 'static/favicon.png')
 @template('index.html')
 async def index(request: web.Request) -> dict:
     db: aiosqlite.Connection = request.app['db']
-    display_names: Dict[str, str] = request.app['display_names']
 
-    async with db.execute(constants.SELECT_LATEST_QUERY) as cursor:  # type: aiosqlite.Cursor
-        updates = [Update.from_row(row) for row in await cursor.fetchall()]
+    updates = await select_latest(db)
     return {
-        'with_names': [update for update in updates if update.key in display_names],
-        'without_names': [update for update in updates if update.key not in display_names],
+        'updates': updates,
     }
 
 
@@ -37,27 +28,7 @@ async def view(request: web.Request) -> dict:
     db: aiosqlite.Connection = request.app['db']
     key: str = request.match_info['key']
 
-    # TODO: refactor.
-    # TODO: select fewer points.
-    since = (datetime.now() - timedelta(days=7)).timestamp() * 1000  # FIXME: `* 1000`.
-    async with db.execute(constants.SELECT_LATEST_BY_KEY_QUERY, [key]) as cursor:  # type: aiosqlite.Cursor
-        update = Update.from_row(await cursor.fetchone())
-    async with db.execute(constants.SELECT_HISTORY_BY_KEY_QUERY, [key, since]) as cursor:  # type: aiosqlite.Cursor
-        values = [
-            (datetime.fromtimestamp(row['timestamp'] / 1000).astimezone(), loads(row['value']))
-            for row in await cursor.fetchall()
-        ]
-
-    # TODO: run in a thread pool executor.
-    line_chart = pygal.Line(show_legend=False, tooltip_border_radius=10)
-    line_chart.x_labels = [timestamp for timestamp, _ in values]
-    line_chart.add('Value', [value for _, value in values])
-
-    # TODO: run in a thread pool executor.
-    return {
-        'update': update,
-        'line_chart': line_chart.render(is_unicode=True),
-    }
+    raise NotImplementedError()
 
 
 @routes.get('/favicon.png')