+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix passing options via environment variables
 - Always start on port `8443`
 - Use [`aiohttp-sse-client`](https://pypi.org/project/aiohttp-sse-client/)
+- #2 #4 #6 Change database schema, rename `ValueKind` and add titles
 
 ### `0.2`